Delegated ploting mechanisms to dedicated classes
